backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
requests>=2.31
pandas>=2.2
numpy>=1.26
supabase>=2.4
python-dotenv>=1.0
pydantic>=2.6
google-generativeai>=0.7
